Added new Controls component
